Move ViriformatLoader outside of import, and add game count
