+++ tests/Lynx.Test/ViriformatLoaderMultiGameTest.cs
@@ -1,4 +1,3 @@
-using Lynx.Import;
 using Lynx.Model;
 using NUnit.Framework;
 using System.Buffers.Binary;
